[Content] grab original "Content Settings"
